Do not handle closed streams
The stream needs to be able to be cleanly closed. The handler will know
when that happens (right now hard codded to answer that all closed streams
are expected).

Nervion spec marked as pending until the stream handler really knows if
the stream was requested to be closed or not.